dev (m) enables staple sequence assignment
staple sequences can be assigned in a similar fashion to the scaffold. they do not change the scaffold sequence.
modification of the staple sequence is now only possible with  length check. for this reason,  the empty sequence placeholder was changed to ? which is not displayed in the GUI.
Also introduces N to indicate intentionally unspecified nucleotides
SD-192, SD-197